add exception handling
fix testcase
